FIX: broken lnurl pay via lnaddress
